# src/main/resources/specifications/nws-openapi.json
  {
    "openapi": "3.0.3",
    "info": {
      "title": "weather.gov API",
      "description": "weather.gov API",
      "version": "2.6.0"
    },
    "servers": [
      {
        "url": "https://api.weather.gov",
        "description": "Production server"
      }
    ],
    "paths": {
      "/alerts": {
        "get": {
          "description": "Returns all alerts",
          "operationId": "alerts_query",
          "parameters": [
            {
              "name": "active",
              "in": "query",
              "description": "List only active alerts (use /alerts/active endpoints instead)",
              "deprecated": true,
              "schema": {
                "type": "boolean"
              }
            },
            {
              "$ref": "#/components/parameters/QueryStartTime"
            },
            {
              "$ref": "#/components/parameters/QueryEndTime"
            },
            {
              "$ref": "#/components/parameters/AlertStatus"
            },
            {
              "$ref": "#/components/parameters/AlertMessageType"
            },
            {
              "$ref": "#/components/parameters/AlertEventName"
            },
            {
              "$ref": "#/components/parameters/AlertCode"
            },
            {
              "$ref": "#/components/parameters/AlertArea"
            },
            {
              "$ref": "#/components/parameters/AlertPoint"
            },
            {
              "$ref": "#/components/parameters/AlertRegion"
            },
            {
              "$ref": "#/components/parameters/AlertRegionType"
            },
            {
              "$ref": "#/components/parameters/AlertZone"
            },
            {
              "$ref": "#/components/parameters/AlertUrgency"
            },
            {
              "$ref": "#/components/parameters/AlertSeverity"
            },
            {
              "$ref": "#/components/parameters/AlertCertainty"
            },
            {
              "$ref": "#/components/parameters/Limit"
            },
            {
              "$ref": "#/components/parameters/PaginationCursor"
            }
          ],
          "responses": {
            "200": {
              "$ref": "#/components/responses/AlertCollection"
            },
            "301": {
              "description": "Certain common queries may be redirected to discrete URLs"
            },
            "default": {
              "$ref": "#/components/responses/Error"
            }
          }
        }
      },
...
      "/glossary": {
        "get": {
          "description": "Returns glossary terms",
          "operationId": "glossary",
          "responses": {
            "200": {
              "description": "A set of glossary terms",
              "headers": {
                "X-Correlation-Id": {
                  "$ref": "#/components/headers/CorrelationId"
                },
                "X-Request-Id": {
                  "$ref": "#/components/headers/RequestId"
                },
                "X-Server-Id": {
                  "$ref": "#/components/headers/ServerId"
                }
              },
              "content": {
                "application/ld+json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "@context": {
                        "$ref": "#/components/schemas/JsonLdContext"
                      },
                      "glossary": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "term": {
                              "type": "string",
                              "description": "The term being defined"
                            },
                            "definition": {
                              "type": "string",
                              "description": "A definition for the term"
                            }
                          }
                        },
                        "description": "A list of glossary terms"
                      }
                    }
                  }
                }
              }
            },
            "default": {
              "$ref": "#/components/responses/Error"
            }
          },
          "parameters": []
        }
      },
...
      "/stations": {
        "get": {
          "description": "Returns a list of observation stations.",
          "operationId": "obs_stations",
          "parameters": [
            {
              "name": "id",
              "in": "query",
              "description": "Filter by observation station ID",
              "style": "form",
              "explode": false,
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            {
              "name": "state",
              "in": "query",
              "description": "Filter by state/marine area code",
              "style": "form",
              "explode": false,
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AreaCode"
                }
              }
            },
            {
              "$ref": "#/components/parameters/Limit"
            },
            {
              "$ref": "#/components/parameters/PaginationCursor"
            },
            {
              "$ref": "#/components/parameters/ObservationsStationFeatureFlags"
            }
          ],
          "responses": {
            "200": {
              "$ref": "#/components/responses/ObservationStationCollection"
            },
            "default": {
              "$ref": "#/components/responses/Error"
            }
          }
        }
      },
...
      "/products": {
        "get": {
          "description": "Returns a list of text products",
          "operationId": "products_query",
          "parameters": [
            {
              "name": "location",
              "in": "query",
              "description": "Location id",
              "style": "form",
              "explode": false,
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            {
              "name": "start",
              "in": "query",
              "description": "Start time",
              "schema": {
                "type": "string",
                "format": "date-time"
              }
            },
            {
              "name": "end",
              "in": "query",
              "description": "End time",
              "schema": {
                "type": "string",
                "format": "date-time"
              }
            },
            {
              "name": "office",
              "in": "query",
              "description": "Issuing office",
              "style": "form",
              "explode": false,
              "schema": {
                "type": "array",
                "items": {
                  "pattern": "^[A-Z]{4}$",
                  "type": "string"
                }
              }
            },
            {
              "name": "wmoid",
              "in": "query",
              "description": "WMO id code",
              "style": "form",
              "explode": false,
              "schema": {
                "type": "array",
                "items": {
                  "pattern": "^[A-Z]{4}\\d{2}$",
                  "type": "string"
                }
              }
            },
            {
              "name": "type",
              "in": "query",
              "description": "Product code",
              "style": "form",
              "explode": false,
              "schema": {
                "type": "array",
                "items": {
                  "pattern": "^\\w{3}$",
                  "type": "string"
                }
              }
            },
            {
              "name": "limit",
              "in": "query",
              "description": "Limit",
              "schema": {
                "maximum": 500,
                "minimum": 1,
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "success",
              "headers": {
                "X-Correlation-Id": {
                  "$ref": "#/components/headers/CorrelationId"
                },
                "X-Request-Id": {
                  "$ref": "#/components/headers/RequestId"
                },
                "X-Server-Id": {
                  "$ref": "#/components/headers/ServerId"
                }
              },
              "content": {
                "application/ld+json": {
                  "schema": {
                    "$ref": "#/components/schemas/TextProductCollection"
                  }
                }
              }
            },
            "default": {
              "$ref": "#/components/responses/Error"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "AlertCollectionGeoJson": {
          "allOf": [
            {
              "$ref": "#/components/schemas/GeoJsonFeatureCollection"
            },
            {
              "type": "object",
              "properties": {
                "features": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "properties": {
                        "$ref": "#/components/schemas/Alert"
                      }
                    }
                  }
                }
              }
            },
            {
              "$ref": "#/components/schemas/AlertCollection"
            }
          ]
        },
        "AlertCollectionJsonLd": {
          "allOf": [
            {
              "$ref": "#/components/schemas/AlertCollection"
            },
            {
              "type": "object",
              "properties": {
                "@context": {
                  "$ref": "#/components/schemas/JsonLdContext"
                },
                "@graph": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Alert"
                  }
                }
              }
            }
          ]
        },
...
        "AlertCertainty": {
          "enum": [
            "Observed",
            "Likely",
            "Possible",
            "Unlikely",
            "Unknown"
          ],
          "type": "string"
        },
...
        "AlertSeverity": {
          "enum": [
            "Extreme",
            "Severe",
            "Moderate",
            "Minor",
            "Unknown"
          ],
          "type": "string"
        },
...
        "AlertUrgency": {
          "enum": [
            "Immediate",
            "Expected",
            "Future",
            "Past",
            "Unknown"
          ],
          "type": "string"
        },
...
        "AlertAtomFeed": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "xml": {
                "namespace": "http://www.w3.org/2005/Atom"
              }
            },
            "generator": {
              "type": "string",
              "xml": {
                "namespace": "http://www.w3.org/2005/Atom"
              }
            },
            "updated": {
              "type": "string",
              "xml": {
                "namespace": "http://www.w3.org/2005/Atom"
              }
            },
            "author": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "xml": {
                    "namespace": "http://www.w3.org/2005/Atom"
                  }
                }
              },
              "xml": {
                "namespace": "http://www.w3.org/2005/Atom"
              }
            },
            "title": {
              "type": "string",
              "xml": {
                "namespace": "http://www.w3.org/2005/Atom"
              }
            },
            "entry": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/AlertAtomEntry"
              },
              "xml": {
                "namespace": "http://www.w3.org/2005/Atom"
              }
            }
          },
          "description": "An alert feed in Atom format",
          "xml": {
            "name": "feed",
            "namespace": "http://www.w3.org/2005/Atom"
          }
        },
        "AreaCode": {
          "oneOf": [
            {
              "$ref": "#/components/schemas/StateTerritoryCode"
            },
            {
              "$ref": "#/components/schemas/MarineAreaCode"
            }
          ],
          "description": "State/territory codes and marine area codes"
        },
...
        "GeoJsonFeatureCollection": {
          "required": [
            "type",
            "features"
          ],
          "type": "object",
          "properties": {
            "@context": {
              "$ref": "#/components/schemas/JsonLdContext"
            },
            "type": {
              "enum": [
                "FeatureCollection"
              ],
              "type": "string"
            },
            "features": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/GeoJsonFeature"
              }
            }
          },
          "description": "A GeoJSON feature collection. Please refer to IETF RFC 7946 for information on the GeoJSON format."
        },
...
        "JsonLdContext": {
          "anyOf": [
            {
              "type": "array",
              "items": {}
            },
            {
              "type": "object"
            }
          ]
        },
...
        "MarineAreaCode": {
          "enum": [
            "AM",
            "AN",
            "GM",
            "LC",
            "LE",
            "LH",
            "LM",
            "LO",
            "LS",
            "PH",
            "PK",
            "PM",
            "PS",
            "PZ",
            "SL"
          ],
          "type": "string",
          "description": "Marine area code as defined in NWS Directive 10-302:\n* AM: Western North Atlantic Ocean and along U.S. East Coast south of Currituck Beach Light NC following the coastline into Gulf of Mexico to Ocean Reef FL including the Caribbean\n* AN: Western North Atlantic Ocean and along U.S. East Coast from Canadian border south to Currituck Beach Light NC\n* GM: Gulf of Mexico and along the U.S. Gulf Coast from the Mexican border to Ocean Reef FL\n* LC: Lake St. Clair\n* LE: Lake Erie\n* LH: Lake Huron\n* LM: Lake Michigan\n* LO: Lake Ontario\n* LS: Lake Superior\n* PH: Central Pacific Ocean including Hawaiian waters\n* PK: North Pacific Ocean near Alaska and along Alaska coastline including the Bering Sea and the Gulf of Alaska\n* PM: Western Pacific Ocean including Mariana Island waters\n* PS: South Central Pacific Ocean including American Samoa waters\n* PZ: Eastern North Pacific Ocean and along U.S. West Coast from Canadian border to Mexican border\n* SL: St. Lawrence River above St. Regis\n"
        },
        "MarineRegionCode": {
          "enum": [
            "AL",
            "AT",
            "GL",
            "GM",
            "PA",
            "PI"
          ],
          "type": "string",
          "description": "Marine region code. These are groups of marine areas combined.\n* AL: Alaska waters (PK)\n* AT: Atlantic Ocean (AM, AN)\n* GL: Great Lakes (LC, LE, LH, LM, LO, LS, SL)\n* GM: Gulf of Mexico (GM)\n* PA: Eastern Pacific Ocean and U.S. West Coast (PZ)\n* PI: Central and Western Pacific (PH, PM, PS)\n"
        },
...
        "NWSZoneID": {
          "pattern": "^(A[KLMNRSZ]|C[AOT]|D[CE]|F[LM]|G[AMU]|I[ADLN]|K[SY]|L[ACEHMOS]|M[ADEHINOPST]|N[CDEHJMVY]|O[HKR]|P[AHKMRSWZ]|S[CDL]|T[NX]|UT|V[AIT]|W[AIVY]|[HR]I)[CZ]\\d{3}$",
          "type": "string",
          "description": "UGC identifier for a NWS forecast zone or county.\nThe first two letters will correspond to either a state code or marine area code (see #/components/schemas/StateTerritoryCode and #/components/schemas/MarineAreaCode for lists of valid letter combinations).\nThe third letter will be Z for public/fire zone or C for county.\n"
        },
...
        "ObservationStationCollectionGeoJson": {
          "allOf": [
            {
              "$ref": "#/components/schemas/GeoJsonFeatureCollection"
            },
            {
              "type": "object",
              "properties": {
                "features": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "properties": {
                        "$ref": "#/components/schemas/ObservationStation"
                      }
                    }
                  }
                },
                "observationStations": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uri"
                  }
                },
                "pagination": {
                  "$ref": "#/components/schemas/PaginationInfo"
                }
              }
            }
          ]
        },
        "ObservationStationCollectionJsonLd": {
          "type": "object",
          "properties": {
            "@context": {
              "$ref": "#/components/schemas/JsonLdContext"
            },
            "@graph": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ObservationStation"
              }
            },
            "observationStations": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uri"
              }
            },
            "pagination": {
              "$ref": "#/components/schemas/PaginationInfo"
            }
          },
          "additionalProperties": false
        },
...
        "PointString": {
          "pattern": "^(-?\\d+(?:\\.\\d+)?),(-?\\d+(?:\\.\\d+)?)$",
          "type": "string"
        },
        "ProblemDetail": {
          "required": [
            "type",
            "title",
            "status",
            "detail",
            "instance",
            "correlationId"
          ],
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "description": "A URI reference (RFC 3986) that identifies the problem type. This is only an identifier and is not necessarily a resolvable URL.\n",
              "format": "uri",
              "default": "about:blank",
              "example": "urn:noaa:nws:api:UnexpectedProblem"
            },
            "title": {
              "type": "string",
              "description": "A short, human-readable summary of the problem type.",
              "example": "Unexpected Problem"
            },
            "status": {
              "maximum": 999,
              "minimum": 100,
              "type": "number",
              "description": "The HTTP status code (RFC 7231, Section 6) generated by the origin server for this occurrence of the problem.\n",
              "example": 500
            },
            "detail": {
              "type": "string",
              "description": "A human-readable explanation specific to this occurrence of the problem.",
              "example": "An unexpected problem has occurred."
            },
            "instance": {
              "type": "string",
              "description": "A URI reference (RFC 3986) that identifies the specific occurrence of the problem. This is only an identifier and is not necessarily a resolvable URL.\n",
              "format": "uri",
              "example": "urn:noaa:nws:api:request:493c3a1d-f87e-407f-ae2c-24483f5aab63"
            },
            "correlationId": {
              "type": "string",
              "description": "A unique identifier for the request, used for NWS debugging purposes. Please include this identifier with any correspondence to help us investigate your issue.\n",
              "example": "493c3a1d-f87e-407f-ae2c-24483f5aab63"
            }
          },
          "description": "Detail about an error. This document conforms to RFC 7807 (Problem Details for HTTP APIs).",
          "additionalProperties": true
        },
...
        "StateTerritoryCode": {
          "enum": [
            "AL",
            "AK",
            "AS",
            "AR",
            "AZ",
            "CA",
            "CO",
            "CT",
            "DE",
            "DC",
            "FL",
            "GA",
            "GU",
            "HI",
            "ID",
            "IL",
            "IN",
            "IA",
            "KS",
            "KY",
            "LA",
            "ME",
            "MD",
            "MA",
            "MI",
            "MN",
            "MS",
            "MO",
            "MT",
            "NE",
            "NV",
            "NH",
            "NJ",
            "NM",
            "NY",
            "NC",
            "ND",
            "OH",
            "OK",
            "OR",
            "PA",
            "PR",
            "RI",
            "SC",
            "SD",
            "TN",
            "TX",
            "UT",
            "VT",
            "VI",
            "VA",
            "WA",
            "WV",
            "WI",
            "WY",
            "MP",
            "PW",
            "FM",
            "MH"
          ],
          "type": "string"
        },
...
        "TextProductCollection": {
          "type": "object",
          "properties": {
            "@context": {
              "$ref": "#/components/schemas/JsonLdContext"
            },
            "@graph": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TextProduct"
              }
            }
          },
          "additionalProperties": false
        },
...
      },
      "responses": {
        "AlertCollection": {
          "description": "A collection of alerts.",
          "headers": {
            "X-Correlation-Id": {
              "$ref": "#/components/headers/CorrelationId"
            },
            "X-Request-Id": {
              "$ref": "#/components/headers/RequestId"
            },
            "X-Server-Id": {
              "$ref": "#/components/headers/ServerId"
            }
          },
          "content": {
            "application/geo+json": {
              "schema": {
                "$ref": "#/components/schemas/AlertCollectionGeoJson"
              }
            },
            "application/ld+json": {
              "schema": {
                "$ref": "#/components/schemas/AlertCollectionJsonLd"
              }
            },
            "application/atom+xml": {
              "schema": {
                "$ref": "#/components/schemas/AlertAtomFeed"
              }
            }
          },
          "x-url-content-negotiation-extensions": {
            "json": "application/geo+json",
            "atom": "application/atom+xml"
          }
        },
        "Error": {
          "description": "An error response.",
          "headers": {
            "X-Correlation-Id": {
              "$ref": "#/components/headers/CorrelationId"
            },
            "X-Request-Id": {
              "$ref": "#/components/headers/RequestId"
            },
            "X-Server-Id": {
              "$ref": "#/components/headers/ServerId"
            }
          },
          "content": {
            "application/problem+json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemDetail"
              }
            }
          }
        },
...
        "ObservationStationCollection": {
          "description": "A collection of observation stations.",
          "headers": {
            "X-Correlation-Id": {
              "$ref": "#/components/headers/CorrelationId"
            },
            "X-Request-Id": {
              "$ref": "#/components/headers/RequestId"
            },
            "X-Server-Id": {
              "$ref": "#/components/headers/ServerId"
            }
          },
          "content": {
            "application/geo+json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationStationCollectionGeoJson"
              }
            },
            "application/ld+json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationStationCollectionJsonLd"
              }
            }
          }
        }
      },
      "parameters": {
        "AlertArea": {
          "name": "area",
          "in": "query",
          "description": "State/territory code or marine area code\nThis parameter is incompatible with the following parameters: point, region, region_type, zone\n",
          "style": "form",
          "explode": false,
          "schema": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AreaCode"
            }
          }
        },
        "AlertCertainty": {
          "name": "certainty",
          "in": "query",
          "description": "Certainty (observed, likely, possible, unlikely, unknown)",
          "style": "form",
          "explode": false,
          "schema": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlertCertainty"
            }
          }
        },
        "AlertCode": {
          "name": "code",
          "in": "query",
          "description": "Event code",
          "style": "form",
          "explode": false,
          "schema": {
            "type": "array",
            "items": {
              "pattern": "^\\w{3}$",
              "type": "string"
            }
          }
        },
        "AlertEventName": {
          "name": "event",
          "in": "query",
          "description": "Event name",
          "style": "form",
          "explode": false,
          "schema": {
            "type": "array",
            "items": {
              "pattern": "^[A-Za-z0-9 ]+$",
              "type": "string"
            }
          }
        },
        "AlertMessageType": {
          "name": "message_type",
          "in": "query",
          "description": "Message type (alert, update, cancel)",
          "style": "form",
          "explode": false,
          "schema": {
            "type": "array",
            "items": {
              "enum": [
                "alert",
                "update",
                "cancel"
              ],
              "type": "string"
            }
          }
        },
        "AlertPoint": {
          "name": "point",
          "in": "query",
          "description": "Point (latitude,longitude)\nThis parameter is incompatible with the following parameters: area, region, region_type, zone\n",
          "schema": {
            "$ref": "#/components/schemas/PointString"
          }
        },
        "AlertRegion": {
          "name": "region",
          "in": "query",
          "description": "Marine region code\nThis parameter is incompatible with the following parameters: area, point, region_type, zone\n",
          "style": "form",
          "explode": false,
          "schema": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MarineRegionCode"
            }
          }
        },
        "AlertRegionType": {
          "name": "region_type",
          "in": "query",
          "description": "Region type (land or marine)\nThis parameter is incompatible with the following parameters: area, point, region, zone\n",
          "schema": {
            "enum": [
              "land",
              "marine"
            ],
            "type": "string"
          }
        },
        "AlertSeverity": {
          "name": "severity",
          "in": "query",
          "description": "Severity (extreme, severe, moderate, minor, unknown)",
          "style": "form",
          "explode": false,
          "schema": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlertSeverity"
            }
          }
        },
        "AlertStatus": {
          "name": "status",
          "in": "query",
          "description": "Status (actual, exercise, system, test, draft)",
          "style": "form",
          "explode": false,
          "schema": {
            "type": "array",
            "items": {
              "enum": [
                "actual",
                "exercise",
                "system",
                "test",
                "draft"
              ],
              "type": "string"
            }
          }
        },
        "AlertUrgency": {
          "name": "urgency",
          "in": "query",
          "description": "Urgency (immediate, expected, future, past, unknown)",
          "style": "form",
          "explode": false,
          "schema": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlertUrgency"
            }
          }
        },
        "AlertZone": {
          "name": "zone",
          "in": "query",
          "description": "Zone ID (forecast or county)\nThis parameter is incompatible with the following parameters: area, point, region, region_type\n",
          "style": "form",
          "explode": false,
          "schema": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NWSZoneID"
            }
          }
        },
...
        "Limit": {
          "name": "limit",
          "in": "query",
          "description": "Limit",
          "schema": {
            "maximum": 500,
            "minimum": 1,
            "type": "integer",
            "default": 500
          }
        },
...
        "ObservationsStationFeatureFlags": {
          "name": "Feature-Flags",
          "in": "header",
          "description": "Enable future and experimental features (see documentation for more info):\n* obs_station_provider: Include the data provider in the metadata about observation stations\n",
          "required": false,
          "style": "simple",
          "explode": false,
          "schema": {
            "type": "array",
            "items": {
              "enum": [
                "obs_station_provider"
              ],
              "type": "string"
            }
          }
        },
...
        "PaginationCursor": {
          "name": "cursor",
          "in": "query",
          "description": "Pagination cursor",
          "schema": {
            "type": "string"
          }
        },
...
        "QueryStartTime": {
          "name": "start",
          "in": "query",
          "description": "Start time",
          "schema": {
            "type": "string",
            "format": "date-time"
          }
        },
        "QueryEndTime": {
          "name": "end",
          "in": "query",
          "description": "End time",
          "schema": {
            "type": "string",
            "format": "date-time"
          }
        },
...
      },
      "headers": {
        "CorrelationId": {
          "description": "A unique identifier for the request, used for NWS debugging purposes. Please include this identifier with any correspondence to help us investigate your issue.\n",
          "schema": {
            "type": "string"
          }
        },
        "RequestId": {
          "description": "A unique identifier for the request, used for NWS debugging purposes. Please include this identifier with any correspondence to help us investigate your issue.\n",
          "schema": {
            "type": "string"
          }
        },
        "ServerId": {
          "description": "The identifier of the server that generated the response, used for NWS debugging purposes. Please include this identifier with any correspondence to help us investigate your issue.\n",
          "schema": {
            "type": "string"
          }
        }
      },
      "securitySchemes": {
        "userAgent": {
          "type": "apiKey",
          "description": "We require that all consumers of the API include a User-Agent header in requests. This is due to a high number of scripts exhibiting abusive behavior (intentional or unintentional). We recommend setting the value to something that identifies your application and includes a contact email. This will help us contact you if we notice unusual behavior and also aid in troubleshooting issues.\nThe API remains open and free to use and there are no limits imposed based on the User-Agent string.\nThis mechanism will be replaced with a more typical API key system at a later date.\n",
          "name": "User-Agent",
          "in": "header"
        }
      }
    },
    "security": [
      {
        "userAgent": []
      }
    ],
...
  }